requirements.txt
Flask>=3.0.0
pandas>=2.1.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
werkzeug>=3.0.0
requests>=2.31.0
xlrd>=2.0.1
